Change 'faraday_middleware' to runtime dependency
In the README, it is documented that this gem needs "faraday_middleware" gem installed.
We should make it a runtime dependency so that people don't have to specify it explicitly, or don't accidentally remove "faraday_middleware"

For the context: in our repo, we have "faraday_middleware" installed because it's the dependency of another gem. When we removed that gem we almost accidentally remove "faraday_middleware". Only upon reading README of "faraday_middleware-aws-sigv4" do we realize it is still needed